Responsywność oraz dodatkowe komponenty
