+++ src/components/Table/Table.module.css
@@ -10,157 +10,26 @@
   margin: auto;
 }
 
-table.table {
-  border-collapse: separate;
-  width: 100%;
-  border-spacing: 1px;
+@media screen and (min-width: 768px) {
+  .table-container {
+    width: 336px;
+  }
 
-  margin: auto;
-  background-color: var(
-    --background-gray
-  ); /*TODO do usuniecia jak bedzie dobry bg color na stronie */
+  .filters {
+    width: 100%;
+    flex-direction: row;
+    justify-content: space-between;
+    gap: 16px;
+  }
 }
+@media screen and (min-width: 1280px) {
+  .table-container {
+    width: 395px;
+  }
 
-.table thead {
-  height: 58px;
-  background-color: var(--white);
-}
+  .filters {
+    width: 100%;
 
-table.table th {
-  border: 0;
-  box-shadow: 0 0 0 1px var(--white);
-}
-
-table.table thead tr:first-child th:first-child {
-  border-top-left-radius: 30px;
-  padding: 15px 0 15px 20px;
-  text-align: left;
-}
-
-table.table thead tr:last-child th:first-child {
-  border-bottom-left-radius: 30px;
-}
-
-table.table thead tr:first-child th:last-child {
-  border-top-right-radius: 30px;
-  padding: 15px 20px 15px 0;
-  text-align: right;
-}
-
-table.table thead tr:last-child th:last-child {
-  border-bottom-right-radius: 30px;
-}
-
-/* .table td {
-  padding: 15px;
-} */
-
-.table .data-row td {
-  position: relative;
-  border-bottom: 1px solid var(--row-divider);
-  box-shadow: 0px 1px 0px 0px #ffffff99;
-}
-
-.table tbody tr:first-child {
-  text-align: left;
-
-  /* line-height: 24px; */
-}
-
-.table tbody tr td:last-child {
-  text-align: right;
-}
-
-.expenses td:last-child {
-  color: var(--pink300);
-}
-.income td:last-child {
-  color: var(--green300);
-}
-
-.table tr td:first-child {
-  padding: 15px 0 15px 20px;
-}
-
-.table tr td:last-child {
-  padding: 15px 20px 15px 0px;
-}
-
-.expenses td:first-child {
-  padding: 15px 0 0 15px;
-}
-
-.expenses td:last-child {
-  padding: 15px 15px 0 0;
-}
-.table .data-row td:first-child {
-  padding: 15px 0 15px 56px;
-}
-
-.data-row td:first-child:before {
-  content: "";
-  display: inline-block;
-  width: 24px;
-  height: 24px;
-  margin-right: 16px;
-  background-color: teal;
-  border-radius: 2px;
-  position: absolute;
-  top: 10px;
-  left: 20px;
-}
-
-.select {
-  border: 1px solid var(--black);
-  border-radius: 30px;
-}
-
-.dropdown {
-  border-radius: 20px;
-  background-color: #ffffffb2;
-  width: 100%;
-  box-shadow: 0px 6px 15px 0px #0000001a;
-
-  overflow: hidden;
-  overflow-y: auto;
-  max-height: 157px;
-
-  backdrop-filter: blur(15px);
-}
-/* 
-.dropdown::-webkit-scrollbar {
-  width: 14px;
-  height: 14px;
-}
-
-.dropdown::-webkit-scrollbar-track {
-  border: rgb(180, 180, 180);
-  background-color: #ff6536;
-}
-
-.dropdown::-webkit-scrollbar-thumb {
-  background-color: #3677ef;
-  border: 1px solid rgb(193, 193, 193);
-} */
-
-.option {
-  height: 28px;
-  padding-left: 20px;
-  vertical-align: middle;
-
-  padding-top: 5px;
-}
-.option:hover {
-  background-color: var(--white);
-}
-.option:first-child {
-  margin-top: 10px;
-}
-
-.valueContainer {
-  padding-left: 20px;
-}
-
-.dropdownIndicator {
-  margin-right: 20px;
+    gap: 32px;
+  }
 }
