+++ src/components/Table/TableStats/TableStats.module.css
@@ -1,0 +1,99 @@
+.table {
+  border-collapse: separate;
+  width: 100%;
+  border-spacing: 1px;
+
+  margin: auto;
+  background-color: var(
+    --background-gray
+  ); /*TODO do usuniecia jak bedzie dobry bg color na stronie */
+}
+
+.table thead {
+  height: 58px;
+  background-color: var(--white);
+}
+
+.table th {
+  border: 0;
+  box-shadow: 0 0 0 1px var(--white);
+}
+
+.table thead tr:first-child th:first-child {
+  border-top-left-radius: 30px;
+  padding: 15px 0 15px 20px;
+  text-align: left;
+}
+
+.table thead tr:last-child th:first-child {
+  border-bottom-left-radius: 30px;
+}
+
+.table thead tr:first-child th:last-child {
+  border-top-right-radius: 30px;
+  padding: 15px 20px 15px 0;
+  text-align: right;
+}
+
+.table thead tr:last-child th:last-child {
+  border-bottom-right-radius: 30px;
+}
+
+/* .table td {
+  padding: 15px;
+} */
+
+.table .data-row td {
+  position: relative;
+  border-bottom: 1px solid var(--row-divider);
+  box-shadow: 0px 1px 0px 0px #ffffff99;
+}
+
+.table tbody tr:first-child {
+  text-align: left;
+
+  /* line-height: 24px; */
+}
+
+.table tbody tr td:last-child {
+  text-align: right;
+}
+
+.expenses td:last-child {
+  color: var(--pink300);
+}
+.income td:last-child {
+  color: var(--green300);
+}
+
+.table tr td:first-child {
+  padding: 15px 0 15px 20px;
+}
+
+.table tr td:last-child {
+  padding: 15px 20px 15px 0px;
+}
+
+.expenses td:first-child {
+  padding: 15px 0 0 15px;
+}
+
+.expenses td:last-child {
+  padding: 15px 15px 0 0;
+}
+.table .data-row td:first-child {
+  padding: 15px 0 15px 56px;
+}
+
+.data-row td:first-child:before {
+  content: "";
+  display: inline-block;
+  width: 24px;
+  height: 24px;
+  margin-right: 16px;
+  background-color: teal;
+  border-radius: 2px;
+  position: absolute;
+  top: 10px;
+  left: 20px;
+}
